replace manual oauth with passport
As it turns out, managing my own oauth, wrapping the library in good Promise-based functions, and managing sessions on my own is a complex, difficult issue, with lots of edge cases and some overhead that I don't need to deal with. So, as much as it pains me, I'm getting rid of my nice functions and converting it over to use passport.