structure data; update state on response
Data is structured on the server. I've leaned towards a thorough trimming of chuft in order to keep down the load on the client. State in MainView is updated on response. The plan is to have a conditional and, if (this.state.data), display a component that displays the results (as opposed to SearchPage)